user can see points on their dashboard
